redesign and new restrictions

+++ rest/PersistanceManager.class.php
@@ -26,56 +26,61 @@ class PersistanceManager{
       $statement = $this->pdo->prepare($query);
       $statement->execute($company);
       $company['id'] = $this->pdo->lastInsertId();
-      Flight::json("User added");
+      return true;
     } else
-        Flight::json("Email exists");
+        return false;
   }
 
   public function add_driver($driver){
     $query = "SELECT * FROM vehicle WHERE companyID = :id AND vehicleID = :vehicleID";
     $statement = $this->pdo->prepare($query);
-    $statement->execute(['id' => $driver["company"], "vehicleID" => $driver["vehicle"]]);
+    $statement->execute(['id' => $driver["companyID"], "vehicleID" => $driver["vehicleID"]]);
     $vehicle = $statement->fetch();
-    if($vehicle != null) {
+    if($vehicle != null && $vehicle["available"] == 1) {
       $query = "INSERT INTO driver
               (firstName,
                lastName,
-               vehicle,
-               company,
-               rating)
+               rating,
+               vehicleID,
+               companyID)
               VALUES (:firstName,
                       :lastName,
-                      :vehicle,
-                      :company,
-                      50)";
+                      50,
+                      :vehicleID,
+                      :companyID)";
       $statement = $this->pdo->prepare($query);
       $statement->execute($driver);
       $driver['id'] = $this->pdo->lastInsertId();
+
+      $query = "UPDATE vehicle SET available = :available WHERE vehicleID = :id";
+      $statement = $this->pdo->prepare($query);
+      $statement->execute(["available" => 0, "id" => $driver["vehicleID"]]);
+
       return 1;
     } else {
       return 0;
     }
+
   }
 
   public function get_drivers($id){
-    $query = "SELECT driverID, firstName, lastName, manufacturer, rating FROM driver JOIN vehicle ON vehicle = vehicleID WHERE company = :id";
-    $statement = $this->pdo->prepare($query);
-    $statement->execute(['id' => $id]);
-    $drivers = $statement->fetchAll();
-    return $drivers;
+    return $this->pdo->query("SELECT driverID, firstName, lastName, rating, manufacturer, model from driver d JOIN vehicle v ON d.vehicleID = v.vehicleID")->fetchAll();
   }
 
   public function get_all_drivers() {
-    return $this->pdo->query("SELECT firstName, lastName, manufacturer, companyName, rating FROM driver d JOIN company as c ON company = companyID JOIN vehicle as v ON vehicle = vehicleID ORDER BY rating DESC LIMIT 10")->fetchAll();
+    return $this->pdo->query("SELECT firstName, lastName, manufacturer, companyName, rating FROM driver d JOIN company as c ON company = companyID JOIN vehicle as v ON vehicleID = vehicleID ORDER BY rating DESC LIMIT 10")->fetchAll();
   }
 
+  public function get_all_companies() {
+    return $this->pdo->query("SELECT companyID, companyName, companyEmail, description FROM company")->fetchAll();
+  }
 
   public function count_all_drivers(){
 
   }
 
   public function get_driver($id){
-    $query = "SELECT driverID, firstName, lastName, vehicle FROM driver WHERE driverID = :id";
+    $query = "SELECT driverID, firstName, lastName, vehicleID FROM driver WHERE driverID = :id";
     $statement = $this->pdo->prepare($query);
     $statement->execute(['id' => $id]);
     $driver = $statement->fetch();
@@ -85,29 +90,43 @@ class PersistanceManager{
   public function update_driver($id, $request){
     $firstName = $request["fname"];
     $lastName = $request["lname"];
-    $vehicle = $request["vehicle"];
-    $query = "UPDATE driver SET firstName = :firstName, lastName = :lastName, vehicle = :vehicle WHERE driverID = :id";
+    $vehicleID = $request["vehicleID"];
+    $query = "UPDATE driver SET firstName = :firstName, lastName = :lastName, vehicleID = :vehicleID WHERE driverID = :id";
     $statement = $this->pdo->prepare($query);
-    $statement->execute(["id" => $id, "firstName" => $firstName, "lastName" => $lastName, "vehicle" => $vehicle]);
+    $statement->execute(["id" => $id, "firstName" => $firstName, "lastName" => $lastName, "vehicleID" => $vehicleID]);
   }
 
   public function delete_driver($id){
+    $query = "Select vehicleID from driver where driverID = :id";
+    $statement = $this->pdo->prepare($query);
+    $statement->execute(["id" => $id]);
+    $vehicle = $statement->fetch();;
+    $query = "UPDATE vehicle SET available = 1 WHERE vehicleID = ?";
+    $statement = $this->pdo->prepare($query);
+    $statement->execute([$vehicle["vehicleID"]]);
+
     $query = "DELETE FROM driver WHERE driverID = ?";
     $statement = $this->pdo->prepare($query);
     $statement->execute([$id]);
   }
 
-
+  public function delete_company($id){
+    $query = "DELETE FROM company WHERE companyID = ?";
+    $statement = $this->pdo->prepare($query);
+    $statement->execute([$id]);
+  }
 
   public function add_vehicle($vehicle){
     $query = "INSERT INTO vehicle
             (manufacturer,
              model,
              year,
+             available,
              companyID)
             VALUES (:manufacturer,
                     :model,
                     :year,
+                    1,
                     :company)";
     $statement = $this->pdo->prepare($query);
     $statement->execute($vehicle);
@@ -116,6 +135,14 @@ class PersistanceManager{
 
   public function get_vehicles($id){
     $query = "SELECT vehicleID, manufacturer, model, year FROM vehicle WHERE companyID = :id";
+    $statement = $this->pdo->prepare($query);
+    $statement->execute(['id' => $id]);
+    $vehicles = $statement->fetchAll();
+    return $vehicles;
+  }
+
+  public function get_available_vehicles($id){
+    $query = "SELECT * FROM vehicle WHERE companyID = :id AND available = 1";
     $statement = $this->pdo->prepare($query);
     $statement->execute(['id' => $id]);
     $vehicles = $statement->fetchAll();
@@ -140,9 +167,17 @@ class PersistanceManager{
   }
 
   public function delete_vehicle($id){
-    $query = "DELETE FROM vehicle WHERE vehicleID = ?";
+    $query = "SELECT * FROM driver WHERE vehicleID = ?";
     $statement = $this->pdo->prepare($query);
     $statement->execute([$id]);
+    if ($statement->fetch() == null) {
+      $query = "DELETE FROM vehicle WHERE vehicleID = ?";
+      $statement = $this->pdo->prepare($query);
+      $statement->execute([$id]);
+      return '0';
+    } else {
+      return '1';
+    }
   }
 
   public function get_company_byID($id) {
@@ -151,6 +186,23 @@ class PersistanceManager{
     $statement->execute(['id' => $id]);
     $company = $statement->fetch();
     return $company;
+  }
+
+  public function get_admin($user) {
+    $stmt = $this->pdo->prepare("SELECT * FROM admin WHERE email = :email;");
+    $stmt->execute(array(
+      "email" => $user["companyEmail"],
+    ));
+    $response = $stmt->fetch();
+    if($response){
+      if ($user["companyPassword"] == $response["password"]) {
+        $response["status"] = "success";
+        return $response;
+      } else {
+        $response["status"] = "fail";
+        return $response;
+      }
+    }
   }
 
   public function get_company($user) {
@@ -173,10 +225,11 @@ class PersistanceManager{
   public function update_company($id, $request){
     $companyName = $request["companyName"];
     $companyEmail = $request["companyEmail"];
+    $description = $request["description"];
     if(!isset($request["companyPassword"]) || $request["companyPassword"] == "") {
-      $query = "UPDATE company SET companyName = :companyName, companyEmail = :companyEmail WHERE companyID = :id";
+      $query = "UPDATE company SET companyName = :companyName, companyEmail = :companyEmail, description = :description WHERE companyID = :id";
       $statement = $this->pdo->prepare($query);
-      $statement->execute(["id" => $id, "companyName" => $companyName, "companyEmail" => $companyEmail]);
+      $statement->execute(["id" => $id, "companyName" => $companyName, "companyEmail" => $companyEmail, "description" => $description]);
     } else {
       $companyPassword = $request["companyPassword"];
       $query = "UPDATE company SET companyName = :companyName, companyEmail = :companyEmail, companyPassword = :companyPassword WHERE companyID = :id";
